requirements.txt
torch>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
transformers>=4.30.0
tokenizers>=0.13.0
tqdm>=4.65.0
wandb>=0.15.0
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0

# Optional dependencies
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# For distributed training
deepspeed>=0.9.0
accelerate>=0.20.0

# For FAISS alternative
lancedb>=0.1.0
